+++ backend/src/main/java/todoktodok/backend/discussion/application/service/query/DiscussionLikeCountDto.java
@@ -10,7 +10,7 @@ public record DiscussionLikeCountDto(
     ) {
         this(
                 discussionId,
-                likeCount.intValue()
+                Math.toIntExact(likeCount)
         );
     }
 }
